Reworked LittleContainerTile

+++ src/main/java/com/creativemd/littletiles/common/gui/SubContainerHammer.java
@@ -124,7 +124,7 @@ public class SubContainerHammer extends SubContainer{
 				else
 					dropstack.getTagCompound().setString("tID", "BlockTileBlock");
 				
-				float missingTiles = blocks-tiles*size.getPercentVolume();
+				double missingTiles = blocks-tiles*size.getPercentVolume();
 				if(missingTiles > 0)
 					ItemTileContainer.addBlock(player, block, meta, missingTiles);
 				//dropstack.stackTagCompound.setString("block", Block.blockRegistry.getNameForObject(block));
